Внес правки путей к картинкам указателей + разделил на функции
