+++ js/util.js
@@ -54,7 +54,34 @@ const getAblePage = function() {
 
 }
 
-export {getRandomNumber, getRandomArrayElement, getRandomArray, fillPhotos, getAblePage};
+const getDisablePage = function () {
+
+  const form = document.querySelector('.ad-form');
+  form.classList.add('ad-form--disabled');
+
+  const filters = document.querySelector('.map__filters');
+  filters.classList.add('map__filters--disabled')
+
+
+  const filter = document.querySelectorAll('select');
+  const inputs = document.querySelectorAll('input');
+  const textArea = document.querySelectorAll('textarea');
+
+  [...filter].forEach((element) => {
+    element.disabled = true;
+  });
+
+  [...inputs].forEach((element) => {
+    element.disabled = true;
+  });
+
+  [...textArea].forEach((element) => {
+    element.disabled = true;
+  });
+
+}
+
+export {getRandomNumber, getRandomArrayElement, getRandomArray, fillPhotos, getAblePage, getDisablePage};
